Replace AnalysisReport with Iterator[Result]
Drop the old AnalysisReport structure in favor of lightweight Result
that just has severity and markdown content.

Switch execute() methods to yield Results which makes the flow a bit
easier to deal with.

Titles are now generated by analyzer.get_title() instead of being
bundled with the reports themselves, and similarly, object_paths
are part of the task_queue metadada rather than Results.

Also added progress bar and removed continuous printing of results
as they become available. We are at a stage where final report
may be good enough.

Logging may still interfere with progress bar though, some other
libraries may solve this, but for now, it's probably okay as-is.